+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **added:** `capacity` field to the `Debug` representation of `Stats` and `Bump(Scope(Guard(Root)))`
+
 ## 0.16.2 (2025-02-21)
 - **changed:** simpler and shorter `Debug` output for `Bump(Scope(Guard(Root)))` and `Chunk(NextIter, PrevIter)`
 - **docs:** removed `rust-patterns` from `package.categories`, added `no-std::no-alloc`